FileChannelSink state is moved to FileChannel, and now only one instance of FileChannelSink is created.
