+++ driver/src/main/java/org/kaazing/k3po/driver/internal/netty/bootstrap/file/FileChannel.java
@@ -16,15 +16,37 @@
 
 package org.kaazing.k3po.driver.internal.netty.bootstrap.file;
 
+import org.jboss.netty.buffer.ChannelBuffer;
+import org.jboss.netty.buffer.ChannelBuffers;
 import org.jboss.netty.channel.ChannelFactory;
+import org.jboss.netty.channel.ChannelHandlerContext;
 import org.jboss.netty.channel.ChannelPipeline;
 import org.jboss.netty.channel.ChannelSink;
+import org.jboss.netty.channel.Channels;
+import org.jboss.netty.channel.MessageEvent;
+import org.jboss.netty.channel.UpstreamMessageEvent;
 import org.kaazing.k3po.driver.internal.netty.bootstrap.channel.AbstractChannel;
 import org.kaazing.k3po.driver.internal.netty.channel.ChannelAddress;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.RandomAccessFile;
+import java.net.URI;
+import java.nio.MappedByteBuffer;
+import java.nio.channels.FileChannel.MapMode;
+import java.util.Map;
+
+import static java.nio.channels.FileChannel.MapMode.READ_ONLY;
+import static java.nio.channels.FileChannel.MapMode.READ_WRITE;
 import static org.jboss.netty.channel.Channels.fireChannelOpen;
 
 public final class FileChannel extends AbstractChannel<FileChannelConfig> {
+
+    // Note that read position and write position are independent in the script. We are keeping
+    // two buffers(view on the same MappedByteBuffer) to achieve this and also don't have to
+    // manage positions explicitly
+    private ChannelBuffer readBuffer;
+    private ChannelBuffer writeBuffer;
 
     FileChannel(ChannelFactory factory, ChannelPipeline pipeline, ChannelSink sink) {
         super(null, factory, pipeline, sink, new DefaultFileChannelConfig());
@@ -44,6 +66,7 @@ public final class FileChannel extends AbstractChannel<FileChannelConfig> {
 
     @Override
     protected boolean setClosed() {
+        // Do we need to call MappedByteBuffer#force() ?
         return super.setClosed();
     }
 
@@ -52,9 +75,90 @@ public final class FileChannel extends AbstractChannel<FileChannelConfig> {
         super.setLocalAddress(localAddress);
     }
 
+    @Override
     public String toString() {
         ChannelAddress localAddress = this.getLocalAddress();
         return localAddress != null ? localAddress.toString() : super.toString();
+    }
+
+    void write(ChannelBuffer channelBuffer) {
+        writeBuffer.writeBytes(channelBuffer);
+    }
+
+    public void setWriteOffset(int offset) {
+        writeBuffer.writerIndex(offset);
+    }
+
+    public void setReadOffset(int offset) {
+        readBuffer.readerIndex(offset);
+    }
+
+    public void fireMessageReceived(ChannelHandlerContext ctx) {
+        Channels.fireMessageReceived(ctx, readBuffer, ctx.getChannel().getRemoteAddress());
+    }
+
+    void fireMessageReceived(FileChannel fileChannel, ChannelAddress fileAddress) {
+        MessageEvent msg = new UpstreamMessageEvent(fileChannel, fileChannel.readBuffer, fileAddress);
+        fileChannel.getPipeline().sendUpstream(msg);
+    }
+
+    void mapFile() throws IOException {
+        Map<String, Object> options = getLocalAddress().getOptions();
+        String mode = (String) options.get("mode");
+        if (mode == null) {
+            mode = "rw";
+        }
+        Long size = (Long) options.get("size");
+        if (size == null) {
+            size = (long) 0;
+        }
+        MappedByteBuffer mappedBuffer = mapFile(getLocalAddress().getLocation(), mode, size);
+        readBuffer = ChannelBuffers.wrappedBuffer(mappedBuffer);
+        setReadOffset(0);
+        writeBuffer = ChannelBuffers.wrappedBuffer(mappedBuffer);
+        setWriteOffset(0);
+    }
+
+    private static MappedByteBuffer mapFile(URI fileAddress, String mode, long size) throws IOException {
+        if (fileAddress.isOpaque()) {
+            // robot scripts specify relative file uri as opaque !
+            URI currentDir = new File(".").toURI();
+            fileAddress = currentDir.resolve(fileAddress.getSchemeSpecificPart());
+        }
+        File location = new File(fileAddress);
+
+        if (!location.exists()) {
+            if (mode.equals("r")) {
+                String msg = String.format("File = %s doesn't exist, cannot be opened in read only mode", location);
+                throw new IllegalArgumentException(msg);
+            }
+            if (size == 0) {
+                String msg = String.format("File = %s is newly created, need size to be specified", location);
+                throw new IllegalArgumentException(msg);
+            }
+        }
+
+        MapMode mapMode;
+        switch (mode) {
+            case "r":
+                mapMode = READ_ONLY;
+                break;
+            case "rw":
+                mapMode = READ_WRITE;
+                break;
+            default:
+                throw new IllegalArgumentException(String.format("Unknown mode = %s for file = %s", mode, location));
+        }
+
+        try (RandomAccessFile file = new RandomAccessFile(location, mode);
+             java.nio.channels.FileChannel channel = file.getChannel()) {
+
+            if (size == 0) {
+                size = channel.size();
+            }
+
+            return channel.map(mapMode, 0, size);
+        }
     }
 
 }
